replaced float literal with constexpr from header

+++ include/consts.h
@@ -1,6 +1,6 @@
 #ifndef CONSTS_H
 #define CONSTS_H
 
-constexpr double EPSILON = 1e-6;
+constexpr float EPSILON = 1e-6;
 
 #endif
